ceph_mon: always try renaming local addrs to id
As it was, we would only attempt renaming 'noname-x' to our id iff
public_addr was not specified. If it so happened to be, we'd still try
to do it, but only as long as the monmap entry had public_addr as its
address.

This would not work with setups such as

  mon host    = 10.10.20.21
  public addr = 172.20.25.21

As the monmap would end up with 10.10.20.21 as its entry's address, and
we would skip the local rename given public_addr is not empty BUT does
not match the monitor's monmap address.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>