mon: MonMap: account for 'public_addr' on build_initial()
Until now, we were ignoring public_addr altogether and relying on 'mon
addr' instead.

'mon addr' should be considered deprecated, as it is an artifact from a
time prior to 'public addr' being a thing.

We will still account for 'mon addr', but as a fallback in case 'public
addr' is not defined.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>