ceph_mon: work around id rename on mkfs for dev clusters
It may not be obvious why this patch is needed, but with the previous
fixes to how we handle entry rename on mkfs we hit a nasty condition:

Say we have `mon.a` running on 127.0.0.1:7117, and want to run `mon.b`
on 127.0.0.1:7118.

Now imagine we run `mon.b`'s mkfs with

  mon_host    = 127.0.0.1:7117
  public_addr = 127.0.0.1:7118

It is expected that `mon.b` ends up with an initial monmap containing

 0: 127.0.0.1:7117  noname-a

And finds its way into quorum, by binding to its public_addr and
attempting communication with `noname-a` on address 127.0.0.1:7117
(which happen to be `mon.a`).

However, what would now actually happen is that we would end up renaming
`noname-a` to `mon.b`, simply because 127.0.0.1:7118 is a local address.
And this would be incorrect.

However, this behavior would go unnoticeable on a real world deployment
(I hope no one deploys multiple monitors on the same host!).

Specifying both monitors' addresses on mon_host, such as,

  mon_host    = 127.0.0.1:7117,127.0.0.1:7118
  public_addr = 127.0.0.1:7118

Would also make the problem go away, but then we would have to change
all the tests exercising this feature: allowing a monitor to find its
way into quorum, even if only a single, remote monitor address is
supplied. And while this behavior should only show up in local dev
clusters, changing the functional tests seems weird, and leaves a nasty
gap in our coverage.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>