mon/MonMap: allow names and addresses alike on mon_host
Previously, we'd only attempt resolving names on `mon host` iff we
failed to parse an address.

This meant that any first pass that contained a name would fail, and
then we'd fall down to parsing all entries as names.

Given we now allow parsing `mon host` as a map, making this work for
both passes seemed silly. Instead, we chose to get rid of the second
pass and fall back immediately to name resolution if *any* entry on `mon
host` happened to fail being parsed.

This now allows us to have

 mon host = 127.0.0.1:7117,b=ceph.com,c=google.com

and end up with

 0: 127.0.0.1:6789/0                   mon.noname-a
 1: 208.113.131.59:6789/0              mon.b         // ceph.com's
 2: 216.58.211.238:6789/0              mon.noname-c  // google.com's
 3: [2a00:1450:4003:805::200e]:6789/0  mon.noname-d  // google.com's

Please note that, should a name resolution return more than one address,
we will make no attempt at using whatever key is provided in the `mon
host` map. We choose to do as such to ensure consistency in the monmap
entries, and not having to decide with one of the addresses will get the
map key, while avoiding having to come up with some clever scheme to
differentiate those addresses.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>