mon: implement cross-service `osd destroy`
We needed to get a bit creative to achieve the `osd destroy` command.

This command will

1. mark the osd as DESTROYED in the osdmap
2. remove all cephx keys (osd and lockbox client)
3. remove the osd's lockbox key

This means we now have a command that spans THREE services. And boy did
we not plan for this to happen.

Therefore, instead of changing the monitor's architecture now, we need
to mock atomicity given we cannot easily guarantee it for a command that
spans multiple services. Instead, we guarantee idempotency for each step.

We manage to chain the steps by relying on callbacks to push the process
one step further each time, until we can finally reply to the client.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>